New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Improvements to Feather GUI.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@22143 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/guitools/Feather/ExternalAxisWidgetBottom.h
@@ -39,7 +39,7 @@ protected:
 	virtual void drawTicks( QPainter* painter, int tickLength);
 	virtual void drawAxisLabel( QPainter* painter );
 private:
-	void drawTick( QPainter* painter, int xPixel, double value, int tickLength);
+	void drawTick( QPainter* painter, double xPixel, double value, int tickLength);
 };
 
 } /* namespace casa */
